task-core-service/requirements.txt
django
djangorestframework
psycopg2-binary
tenacity
pybreaker
tenacity
requests
kafka-python